Ensure that both branches of min() and max() resolve to the same type
This prepares the code to the c++ conversion

+++ src/char/int_party.c
@@ -320,7 +320,7 @@ static struct party_data *inter_party_fromsql(int party_id)
 		return NULL;
 
 	p->party.party_id = party_id;
-	SQL->GetData(inter->sql_handle, 1, &data, &len); memcpy(p->party.name, data, min(len, NAME_LENGTH));
+	SQL->GetData(inter->sql_handle, 1, &data, &len); memcpy(p->party.name, data, min(len, (size_t)NAME_LENGTH));
 	SQL->GetData(inter->sql_handle, 2, &data, NULL); p->party.exp = (atoi(data) ? 1 : 0);
 	SQL->GetData(inter->sql_handle, 3, &data, NULL); p->party.item = atoi(data);
 	SQL->GetData(inter->sql_handle, 4, &data, NULL); leader_id = atoi(data);
@@ -338,7 +338,7 @@ static struct party_data *inter_party_fromsql(int party_id)
 		m = &p->party.member[i];
 		SQL->GetData(inter->sql_handle, 0, &data, NULL); m->account_id = atoi(data);
 		SQL->GetData(inter->sql_handle, 1, &data, NULL); m->char_id = atoi(data);
-		SQL->GetData(inter->sql_handle, 2, &data, &len); memcpy(m->name, data, min(len, NAME_LENGTH));
+		SQL->GetData(inter->sql_handle, 2, &data, &len); memcpy(m->name, data, min(len, (size_t)NAME_LENGTH));
 		SQL->GetData(inter->sql_handle, 3, &data, NULL); m->lv = atoi(data);
 		SQL->GetData(inter->sql_handle, 4, &data, NULL); m->map = mapindex->name2id(data);
 		SQL->GetData(inter->sql_handle, 5, &data, NULL); m->online = (atoi(data) ? 1 : 0);
